% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration}
\alias{calibration}
\title{In-sample calibration graph for binary outcome models}
\usage{
calibration(model, ns = 10)
}
\arguments{
\item{model}{model}

\item{ns}{numver of bins}
}
\value{
\code{ggplot2} graph
}
\description{
In-sample calibration graph for binary outcome models
}
