% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edv.fgls.R
\name{edv.fgls}
\alias{edv.fgls}
\title{FGLS model for estimated dependent variable}
\usage{
edv.fgls(ols, DVvariance = NULL, n = NULL)
}
\arguments{
\item{ols}{Initial OLS model (lm in R)}

\item{DVvariance}{Numeric vector of variances of each y.}

\item{n}{Numeric vector of number of observations (n) on which each y was estimated. The length should be equal the number of observations in the initial model.}
}
\value{
re-estimated model
}
\description{
FGLS model for estimated dependent variable
}
\details{
The function is based completely on the following paper: Lewis, J. B., & Linzer, D. A. (2005). Estimating regression models in which the dependent variable is based on estimates. \strong{Political analysis}, 13(4), 345-364.
There are two possible strategies. If variance of each y is unknown, then Numeric vector of number of observations (n) on which each y was estimated should be provided. For example, y - mean of housholds income, then \code{n} is the number of housholds.
The second possible situation is when a variance of each y is known. In that case, you should provide \code{DVvariance} arguement in the function with numeric vector of such variances (as in case with the first situation with number of observations)
}
