% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldclass.R
\name{goldclass}
\alias{goldclass}
\title{Goldstone claasification of regime types based on Polity-V variables}
\usage{
goldclass(exrec, parcomp)
}
\arguments{
\item{exrec}{array with EXREC variable from Polity-V project}

\item{parcomp}{array with PARCOMP variable from Polity-V project}
}
\value{
array with Goldstone classification:
1 = full autocracy
2 = partial autocracy
3 = partial democracy
4 = partial democracy with factionalism
5 = full democracy
}
\description{
Goldstone claasification of regime types based on Polity-V variables
}
\examples{
polity5data$goldstone_regime <- goldclass(exrec = polity5data$exrec, parcomp = polity5data$parcomp)

}
\references{
Goldstone, J.A., Bates, R.H., Epstein, D.L., Gurr, T.R., Lustik, M.B., Marshall, M.G., Ulfelder, J., Woodward, M., 2010. A Global Model for Forecasting Political Instability. American Journal of Political Science 54, 190–208.
}
