% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdiff.R
\name{sdiff}
\alias{sdiff}
\title{Nonparametric test for comparing means in 2 groups}
\usage{
sdiff(x, y, n = 1000)
}
\arguments{
\item{x}{Numeric array with observations from the first group.}

\item{y}{Numeric array with observations from the second group.}

\item{n}{Number of simulations for constriction Null hypothesis area. Default \code{n=1000}}
}
\value{
summary and plot
}
\description{
Nonparametric test for comparing means in 2 groups
}
\details{
This is nonparametric test (alternative to t-test) based on Monte Carlo Simulations. The idea is straightforward. First, the initial difference between the groups is counted. Null hypothesis - this difference is equal to zero, alternative hypothesis - the difference is significantly different from zero. To test this hypothesis, an iterative procedure is created where values from both groups are randomly mixed (i.e. each observation can be randomly assigned to one of the groups). Due to randomness, the difference in the mean of the mixed groups should be zero. After a thousand such procedures, we construct a distribution of the resulting differences and then simply compare it with the observed difference.
}
\examples{
x = rnorm(1000, 10, 10)
y = rnorm(1000, 11, 2)
sdiff(x, y, n = 1000)

}
