% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_shape_test.R
\name{U_shape_test}
\alias{U_shape_test}
\title{U-shape test for rare events logistic regression}
\usage{
U_shape_test(
  data,
  dep_var,
  ind_var,
  control_vars,
  vcov_type = "HC",
  boot = FALSE,
  n = 1000,
  save_plot = FALSE,
  save_table = FALSE
)
}
\arguments{
\item{data}{Dataframe}

\item{dep_var}{Character. Dependent variable.}

\item{ind_var}{Character. Independent variable.}

\item{control_vars}{Vector of characters with names of control variables.}

\item{vcov_type}{Character. Type from lmtest::coeftest.}

\item{boot}{Logical. Should middle point CI be calculated via bootstrap?}

\item{n}{Int. Number of bootstraps.}

\item{save_plot}{Logical. Should plot be saved.}

\item{save_table}{Logical. Should regression table be saved.}
}
\value{
Plot and regression table with U-shape test summary.
}
\description{
U-shape test for rare events logistic regression
}
\examples{
data("datex")
U_shape_test(data = datex,
dep_var = "NVC_1.3_NONVIOL",
ind_var = "VDEM_v2x_polyarchy_lag",
control_vars = c("UN_Total_Population_log", "UN_Median_Age"),
boot = FALSE,
vcov_type = "HC",
n = 1000,
save_plot = FALSE, save_table = FALSE)

}
