% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datex}
\alias{datex}
\title{Data of revolutionary campaigns and democracy index, 1901-2019}
\format{
\subsection{\code{datex} A data frame with 9,836 rows and 6 columns:}{

\describe{
\item{iso3}{Country code iso3c}
\item{year}{Year}
\item{NVC_1.3_NONVIOL}{Year}
\item{NVC_1.3_VIOL}{Year}
\item{VDEM_v2x_polyarchy}{Electoral democracy index \link{0,1}}
\item{VDEM_v2x_polyarchy_lag}{{Electoral democracy index \link{0,1} at time t-1 for country i}
}
}
}
\source{
NAVCO 1.3, VDEM
}
\usage{
datex
}
\description{
A data from NANVCO 1.3 and VDEM
}
\keyword{datasets}
