% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datex}
\alias{datex}
\title{Data of revolutionary campaigns and democracy index, 1901-2019}
\format{
\subsection{\code{datex} A data frame with 9,836 rows and 6 columns:}{

\describe{
\item{iso3}{Country code in "iso3c" \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}{format}}
\item{year}{Year}
\item{NVC_1.3_NONVIOL}{Binary indicator. Do protestors primary use nonviolent tactic? 1 = Yes}
\item{NVC_1.3_VIOL}{Binary indicator. Do protestors primary use violent tactic? 1 = Yes}
\item{VDEM_v2x_polyarchy}{Electoral democracy index (0,1) from V-Dem project}
\item{VDEM_v2x_polyarchy_lag}{Electoral democracy index (0,1) at time t-1 for country i}
}
}
}
\source{
\href{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/ON9XND}{NAVCO 1.3}, \href{https://v-dem.net/data/the-v-dem-dataset/}{VDEM}
}
\usage{
datex
}
\description{
Data of revolutionary campaigns and democracy index, 1901-2019
}
\keyword{datasets}
