% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_shape_test.R
\name{U_shape_test}
\alias{U_shape_test}
\title{U-shape test for rare events logistic regression}
\usage{
U_shape_test(
  data,
  dep,
  ind,
  cnt,
  mod = "brglm",
  boot = c("none", "nonparam", "factor"),
  factor = NULL,
  n = 1000,
  HC = T,
  const = T,
  table_remove = NULL,
  plot = F,
  tab_save = F
)
}
\arguments{
\item{data}{Dataframe}

\item{dep}{Character. Dependent variable.}

\item{ind}{Character. Independent variable.}

\item{cnt}{Vector of characters with names of control variables.}

\item{mod}{Character. Type of model (now just \code{brglm} is supported)}

\item{boot}{Character. Should middle point and "lines" be calculated via bootstrap? If no - "none", if yes - one can choose bootstrap type}

\item{factor}{Character. Only if \code{boot = "factor"}.}

\item{n}{Number of bootstrap draws}

\item{HC}{Logical. Should heteroscedasticity-consistent SE be used?}

\item{const}{Logical. Should constant be included in the models?}

\item{table_remove}{Vector of variables names that should be excluded from the regression table.}

\item{plot}{Logical. Should plot be depicted?}

\item{tab_save}{Logical. Should table be saved? If yes - table is saved in working directory.}
}
\value{
Plot and regression table with u-shape test summary.
}
\description{
U-shape test for rare events logistic regression
}
\references{
Lind, J. T., & Mehlum, H. (2010). With or without U? The appropriate test for a U‐shaped relationship. \emph{Oxford bulletin of economics and statistics}, \emph{72(1)}, 109-118.

Simonsohn, U. (2018). Two lines: A valid alternative to the invalid testing of U-shaped relationships with quadratic regressions. \emph{Advances in Methods and Practices in Psychological Science}, \emph{1(4)}, 538-555.
}
